Larger Levels v95 ready
Updated to RogueLibs v3
Patcher-Logic has been relocated to Patcher-Class
